Restructure: move source files to src/ folder, add configs and metadata

+++ src/view_segmentations.py
@@ -21,21 +21,22 @@ def load_image_and_annotation(dicom_path, annotation_path):
             annotation = annotation[:, :, 0]
     return image, annotation
 
-def plot_patient_images(row):
+def plot_patient_images(row, save=False):
     """Plot images and annotations for one patient row."""
     patient_class = row['patientclass']
     lesion_count = row['lesion_count']
     ID = row["record_id"]
     path = f'images/patient_{ID}.png'
-    if os.path.exists(path):
+    if os.path.exists(path) and save:
         return  # skip if already exists
     views = ['CC_L', 'MLO_L', 'CC_R', 'MLO_R']
     
-    annotations_present = any(row.get(f'annotation_path_{view}', None) is not None for view in views)
-    if lesion_count == 0 and annotations_present:
-        print(f"!!! Warning: Patient {ID} has lesion_count=0 but annotations exist!")
-    elif lesion_count >= 1 and not annotations_present:
-        print(f"!!! Warning: Patient {ID} has lesion_count={lesion_count} but no annotations!")
+    ## maybe there are no json files
+    # annotations_present = any(row.get(f'annotation_path_{view}', None) is not None for view in views)
+    # if lesion_count == 0 and annotations_present:
+    #     print(f"!!! Warning: Patient {ID} has lesion_count=0 but annotations exist!")
+    # elif lesion_count >= 1 and not annotations_present:
+    #     print(f"!!! Warning: Patient {ID} has lesion_count={lesion_count} but no annotations!")
 
     fig, axes = plt.subplots(2, 4, figsize=(16, 8))
     fig.suptitle(f'ID: {ID} Patient Class: {int(patient_class)} (n={lesion_count})', fontsize=16)
@@ -64,9 +65,9 @@ def plot_patient_images(row):
 
     plt.tight_layout(rect=[0, 0, 1, 0.95])
     plt.show()
-    plt.savefig(path)
+    if os.path.exists(path) and save:
+        plt.savefig(path)
     plt.close()
-
 def main():
     os.chdir('/users/project1/pt01190/EUCAIM-PG-GUM/code')
 
@@ -78,3 +79,4 @@ def main():
 
 if __name__ == '__main__':
     main()
+# %%
